selftests: mlxsw: Adding sanity testing for devlink KVD resource
Add a selftest that can be used to perform basic sanity of the devlink
resource API as well as test the behavior of KVD manipulation in the
driver.

This is a first case of a HW-only test - in order to test the devlink
resource a driver capable of exposing resources has to be provided
first, and KVD [as the only root-resource available at the time] is
the only available option.

This adds 2 additional libraries:
  - devlink_lib.sh - meant to contain wrappers to devlink functionality
    which is agnostic to the underlying device; May eventually relocate
    into another library.
  - devlink_lib_spectrum.sh - spectrum-oriented functionality [e.g.,
    re-partitioning the various KVD sub-parts] would be located here

[1] - There's an issue required debugging when doing this where
linear == 0, so we can't actually maximize anything but the linear;
But 'almost' maximizing the others work.

Signed-off-by: Yuval Mintz <yuvalm@mellanox.com>