mlxsw: spectrum: Add support for auto-negotiation disable mode
Port type and speed (PTYS) register have an_disable_admin field, which
indicates if speed negotiation frames would be send by the port or not.
The problem was that, if a user disabled AN (auto-negotiation) via ethtool
the device was still generating AN frames with the forced speed. Now if
user sets 'autoneg off' AN frames won't be send. Note that for SwitchX2
the behavior doesn't change (i.e support only AN enabled with forced
speed).

Signed-off-by: Tal Bar <talb@mellanox.com>